추가 | 적 피격

+++ Assets/02.Scripts/WeaponData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class WeaponData : ScriptableObject
 {
     public string weaponName;       //이름
-    public int damage;              //데미지
+    public float damage;              //데미지
     public int magazine;            //탄창 수
     public float effectiveRange;    //사정거리
     public float reloadTime;        //장전시간
